refactor: move detail as single file
